+++ changelog.md
@@ -1,6 +1,8 @@
 ## v1.4.1
 
+* Fix the Audio Device and Receive Invites From settings not saving when exiting the game
 * Fix room list not starting at the top
+
 
 ## v1.4.0